refactorred structures, dividing LibInst api receiver to /service, made new generator struct for abstracting domain generation
