+++ api/conn/connect.go
@@ -1,17 +1,8 @@
 package conn
 
-import (
-	"libvirt.org/go/libvirt"
-)
 
 
-func (i *InstHandler)LibvirtConnection(){
-	libvirtInst, err := libvirt.NewConnect("qemu:///system")
-		if err != nil {
-			panic(err)
-}
-	i.LibvirtInst = libvirtInst
-}
+
