+++ api/conn/create.go
@@ -1,84 +1,41 @@
 package conn
 
 import (
-	"bytes"
-	"encoding/json"
 	"fmt"
 	"log"
-	"net/http"
 	"os"
 	"os/exec"
 	"path/filepath"
 
 	"github.com/easy-cloud-Knet/KWS_Core.git/api/parsor"
-	"gopkg.in/yaml.v3"
-	"libvirt.org/go/libvirt"
 )
 
+func (DGL DomainGeneratorLocal) CreateFolder()error{
 
-
-func (i *InstHandler) CreateVM(w http.ResponseWriter, r *http.Request) {
-	var param parsor.VM_Init_Info
-
-
-	if err := json.NewDecoder(r.Body).Decode(&param); err != nil {
-		http.Error(w, "Invalid request payload", http.StatusBadRequest)
-		log.Printf("Error decoding JSON: %v", err)
-		return
-	}
-
-	dirPath := fmt.Sprintf("/var/lib/kws/%s", param.UUID)
+	dirPath := fmt.Sprintf("/var/lib/kws/%s", DGL.DomainStatusManager.UUID)
 	if err := os.MkdirAll(dirPath, 0755); err != nil {
-		http.Error(w, "Failed to create directory", http.StatusInternalServerError)
-		log.Printf("Error creating directory %s: %v", dirPath, err)
-		return
+		
+		return err
 	}
+	return nil
+}
 
+
+func (DGL DomainGeneratorLocal) CloudInitConf(param *parsor.VM_Init_Info)error{
+	DGL.DataParsor.YamlParsor.Parse_data(param)
 	
-	parsedXML:= parsor.XML_Parsor(&param)
-	var parsedUserYaml parsor.User_data_yaml
-	var parsedMetaYaml parsor.Meta_data_yaml
-	parsedUserYaml.Parse_data(&param)
-	parsedMetaYaml.Parse_data(&param)
-
-	// Marshal YAML data
-	marshalledUserData, err := yaml.Marshal(parsedUserYaml)
-	if err != nil {
-		http.Error(w, "Failed to marshal user data", http.StatusInternalServerError)
-		log.Printf("Error marshaling user data: %v", err)
-		return
+	dirPath := fmt.Sprintf("/var/lib/kws/%s", DGL.DomainStatusManager.UUID)
+	if err:= DGL.DataParsor.YamlParsor.FileConfig(dirPath); err!=nil{
+		return err
 	}
 
-	marshalledMetaData, err := yaml.Marshal(parsedMetaYaml)
-	if err != nil {
-		http.Error(w, "Failed to marshal meta data", http.StatusInternalServerError)
-		log.Printf("Error marshaling meta data: %v", err)
-		return
-	}
-	fmt.Println(parsedUserYaml)
+	return nil
+}
 
-	// Write user-data file
-	userConfig := bytes.Buffer{}
-	userConfig.WriteString("#cloud-config\n")
-	userConfig.Write(marshalledUserData)
-	if err := os.WriteFile(filepath.Join(dirPath, "user-data"), userConfig.Bytes(), 0644); err != nil {
-		http.Error(w, "Failed to write user-data file", http.StatusInternalServerError)
-		log.Printf("Error writing user-data file: %v", err)
-		return
-	}
-
-	// Write meta-data file
-	metaConfig := bytes.Buffer{}
-	metaConfig.Write(marshalledMetaData)
-	if err := os.WriteFile(filepath.Join(dirPath, "meta-data"), metaConfig.Bytes(), 0644); err != nil {
-		http.Error(w, "Failed to write meta-data file", http.StatusInternalServerError)
-		log.Printf("Error writing meta-data file: %v", err)
-		return
-	}
-
-	// Execute qemu-img create command
-	baseImage := fmt.Sprintf("/var/lib/kws/baseimg/%s", param.OS )
-	targetImage := filepath.Join(dirPath, fmt.Sprintf("%s.qcow2", param.UUID))
+func (DGL DomainGeneratorLocal) CreateDiskImage() error{
+	dirPath := fmt.Sprintf("/var/lib/kws/%s", DGL.DomainStatusManager.UUID)
+	baseImage := fmt.Sprintf("/var/lib/kws/baseimg/%s", DGL.OS )
+	targetImage := filepath.Join(dirPath, fmt.Sprintf("%s.qcow2", DGL.DomainStatusManager.UUID))
 	qemuImgCmd := exec.Command("qemu-img", "create",
 		"-b", baseImage,
 		"-f", "qcow2",
@@ -91,12 +48,15 @@ func (i *InstHandler) CreateVM(w http.ResponseWriter, r *http.Request) {
 
 	log.Println("Creating disk image...")
 	if err := qemuImgCmd.Run(); err != nil {
-		http.Error(w, "Failed to create disk image", http.StatusInternalServerError)
 		log.Printf("qemu-img command failed: %v", err)
-		return
+		return err
 	}
+	return nil
+}
 
-	// Execute genisoimage command
+func (DGL DomainGeneratorLocal) CreateISOFile()error{
+	dirPath := fmt.Sprintf("/var/lib/kws/%s", DGL.DomainStatusManager.UUID)
+
 	isoOutput := filepath.Join(dirPath, "cidata.iso")
 	userDataPath := filepath.Join(dirPath, "user-data")
 	metaDataPath := filepath.Join(dirPath, "meta-data")
@@ -113,35 +73,8 @@ func (i *InstHandler) CreateVM(w http.ResponseWriter, r *http.Request) {
 
 	log.Println("Generating ISO image...")
 	if err := genisoCmd.Run(); err != nil {
-		http.Error(w, "Failed to generate ISO image", http.StatusInternalServerError)
 		log.Printf("genisoimage command failed: %v", err)
-		return
+		return err
 	}
-
-	dom , err := i.CreateDomainWithXML(parsedXML)
-	if err!= nil{
-		http.Error(w, "faild creating vm", http.StatusConflict)
-	}
-	err = dom.Create()
-	if err!= nil{
-		http.Error(w, "faild starting vm", http.StatusConflict)
-	}
-	fmt.Println(dom)
-	domainInfo,_:= dom.GetInfo()
-	fmt.Println(domainInfo)
-	w.WriteHeader(http.StatusOK)
-	fmt.Fprintf(w, "VM with UUID %s created successfully.", param.UUID)
-}
-
-
-
-func (i *InstHandler) CreateDomainWithXML(config []byte) (*libvirt.Domain, error) {
- 
-	// DomainCreateXMLWithFiles를 호출하여 도메인을 생성합니다.
-	domain, err := i.LibvirtInst.DomainDefineXML(string(config))
-	if err != nil {
-		log.Fatal(err)
-	}
-  return domain ,err
-
+	return nil
 }
