+++ api/conn/delete.go
@@ -1,10 +1,8 @@
 package conn
 
 import (
-	"encoding/json"
 	"fmt"
 	"log"
-	"net/http"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -28,77 +26,12 @@ func DomainDeleterFactory(DomainSeek DomainSeeker, DelType DomainDeleteType, uui
 		},}, nil
 }
 
-func (i *InstHandler)ForceShutDownVM(w http.ResponseWriter, r *http.Request){
-	var param DeleteDomain
-	if err:= json.NewDecoder(r.Body).Decode(&param); err!=nil{
-		http.Error(w, "error decoding body", 1)
-	}
-	DomainSeeker:=&DomainSeekingByUUID{
-		LibvirtInst: i.LibvirtInst,
-		UUID: param.UUID,
-		Domain: make([]*Domain, 0,1),
-	}
-	DomainTerminator,_:= DomainTerminatorFactory(DomainSeeker)
-	domainInfo,err:=DomainTerminator.ShutDownDomain()
-	if err!= nil{
-		http.Error(w,"error while shutting down Domain", http.StatusBadRequest)
-		fmt.Println(err)
-	}
-	//uuid unparsing 중 에러, destroyDom 에서 에러
-	// 수신시 에러 발생 가능 ,추후 에러 핸들링 
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-	response := map[string]interface{}{
-		"message":   fmt.Sprintf("VM with UUID %s Shutdown successfully.", param.UUID),
-		"domainInfo": domainInfo,
-	}
-	
-	if err := json.NewEncoder(w).Encode(response); err != nil {
-		http.Error(w, "Failed to encode response", http.StatusInternalServerError)
-		log.Printf("Error encoding response: %v", err)
-	}
-}
-
-func (i *InstHandler)DeleteVM(w http.ResponseWriter, r *http.Request){
-	var param DeleteDomain
-	if err:= json.NewDecoder(r.Body).Decode(&param);err != nil{
-		http.Error(w, "invalid parameter", http.StatusBadRequest)
-	}
-	DomainSeeker:= &DomainSeekingByUUID{
-		LibvirtInst: i.LibvirtInst,
-		UUID: param.UUID,
-		Domain: make([]*Domain, 0,1),
-	}
-	DomainDeleter,_:=DomainDeleterFactory(DomainSeeker, param.DeletionType, param.UUID)
-	
-	domainInfo, err:=DomainDeleter.DeleteDomain()
-	if err!=nil{
-		fmt.Println(err)
-		http.Error(w,"error while destroying vm", http.StatusInternalServerError)
-	}
-		//uuid unparsing 중 에러, undefine,destroyDom 에서 에러, 켜져 있지만 softdelete가 
-	// 수신시 에러 발생 가능 ,추후 에러 핸들링 
-
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-	
-	response := map[string]interface{}{
-		"message":   fmt.Sprintf("VM with UUID %s Deletion successfully.", param.UUID),
-		"domainInfo": domainInfo,
-	}	
-	if err := json.NewEncoder(w).Encode(response); err != nil {
-		http.Error(w, "Failed to encode response", http.StatusInternalServerError)
-		log.Printf("Error encoding response: %v", err)
-	}
-	
-}
-
 
 func (DD *DomainDeleter) DeleteDomain() (*libvirt.DomainInfo, error){
 	if err :=DD.DomainSeeker.SetDomain(); err!=nil{
 		return &libvirt.DomainInfo{},err
 	}
-	dom,_ := DD.DomainSeeker.returnDomain()
+	dom,_ := DD.DomainSeeker.ReturnDomain()
 
 	isRunning, _ :=dom[0].Domain.IsActive()
 	if isRunning&& DD.DeletionType==SoftDelete {
@@ -129,7 +62,6 @@ func (DD *DomainDeleter) DeleteDomain() (*libvirt.DomainInfo, error){
 	}
 	defer dom[0].Domain.Free()
 	
-	
 	if err!=nil{
 		fmt.Println(err)	
 		//error handler needed
@@ -143,7 +75,7 @@ func (DD *DomainTerminator) ShutDownDomain() (*libvirt.DomainInfo,error){
 	if err :=DD.DomainSeeker.SetDomain(); err!=nil{
 		return &libvirt.DomainInfo{},err
 	}
-	dom,_ := DD.DomainSeeker.returnDomain()
+	dom,_ := DD.DomainSeeker.ReturnDomain()
 	isRunning, _ :=dom[0].Domain.IsActive()
 	if !isRunning {
 		return &libvirt.DomainInfo{},fmt.Errorf("Domain Is Not Running %w", nil)
