+++ api/service/Connection.go
@@ -1,0 +1,20 @@
+package service
+
+import (
+	"libvirt.org/go/libvirt"
+)
+
+
+ 
+
+func (i *InstHandler)LibvirtConnection(){
+	libvirtInst, err := libvirt.NewConnect("qemu:///system")
+		if err != nil {
+			panic(err)
+}
+	i.LibvirtInst = libvirtInst
+}
+
+
+ 
+
