+++ main.go
@@ -5,16 +5,15 @@ import (
 	"fmt"
 	_ "log"
 
-	"github.com/easy-cloud-Knet/KWS_Core.git/api/conn"
+	"github.com/easy-cloud-Knet/KWS_Core.git/api/service"
 	"github.com/easy-cloud-Knet/KWS_Core.git/server"
 )
 
 func main() {
-	var libvirtInst conn.InstHandler
-	var domain conn.Domain
+	var libvirtInst service.InstHandler
 	libvirtInst.LibvirtConnection()
 
-	go server.InitServer(8080, &libvirtInst, &domain)
+	go server.InitServer(8080, &libvirtInst)
 	fmt.Println("working")	
 
 	defer  libvirtInst.LibvirtInst.Close()
