+++ server/server.go
@@ -5,11 +5,11 @@ import (
 	"net/http"
 	"strconv"
 
-	"github.com/easy-cloud-Knet/KWS_Core.git/api/conn"
+	"github.com/easy-cloud-Knet/KWS_Core.git/api/service"
 )
 
 
-func InitServer(portNum int, libvirtInst *conn.InstHandler, domain *conn.Domain){
+func InitServer(portNum int, libvirtInst *service.InstHandler){
 	
 	http.HandleFunc("/getStatus", libvirtInst.ReturnDomainByStatus) //get
 	http.HandleFunc("/createVM", libvirtInst.CreateVM) //post
